Add "sale_partner_bank" module.
This modules adds a 'Bank account' field on partners and sale orders. This field is copied from partner to sale order and then from sale order to customer invoice.

This module is similar to the 'sale_payment' module ; the main difference is that it doesn't have any extra-dependancy i.e. it doesn't depend on the 'account_payment_extension' module.